修改norm下样例代码问题



# message auto-generated for no-merge-commit merge:
!166 merge master into master

修改norm下样例代码问题

Created-by: photoker_bao
Commit-by: photoker_bao
Merged-by: turing_project1
Description: ## 描述
修改norm下deep_norm和deep_norm_grad的example代码问题
验证结果正常：
![image.png](https://raw.gitcode.com/user-images/assets/7665709/b50b5941-689c-418c-aafe-5c8f83d3015c/image.png 'image.png')
![image.png](https://raw.gitcode.com/user-images/assets/7665709/d9891bbe-8997-404f-a5ee-f7673a8b189c/image.png 'image.png')

## 关联的Issue
关联Issue #64

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [ ] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：




See merge request: cann/ops-nn!166

+++ norm/deep_norm_grad/examples/test_aclnn_deep_norm_grad.cpp
@@ -116,15 +116,15 @@ int main()
     aclTensor* outputpdgamma = nullptr;
 
     std::vector<float> dyHostData = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12};
-    std::vector<int32_t> xHostData = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12};
-    std::vector<int32_t> gxHostData = {2, 2, 2, 4, 4, 4, 6, 6, 6, 8, 8, 8};
-    std::vector<int32_t> gammaHostData = {0, 1, 2, 3};
-    std::vector<int32_t> meanHostData = {0, 1, 2};
-    std::vector<int32_t> rstdHostData = {0, 1, 2};
-    std::vector<int32_t> outputpdxHostData = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12};
-    std::vector<int32_t> outputpdgxHostData = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12};
-    std::vector<int32_t> outputpdbetaHostData = {0, 1, 2, 3};
-    std::vector<int32_t> outputpdgammaHostData = {0, 1, 2, 3};
+    std::vector<float> xHostData = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12};
+    std::vector<float> gxHostData = {2, 2, 2, 4, 4, 4, 6, 6, 6, 8, 8, 8};
+    std::vector<float> gammaHostData = {0, 1, 2, 3};
+    std::vector<float> meanHostData = {0, 1, 2};
+    std::vector<float> rstdHostData = {0, 1, 2};
+    std::vector<float> outputpdxHostData = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12};
+    std::vector<float> outputpdgxHostData = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12};
+    std::vector<float> outputpdbetaHostData = {0, 1, 2, 3};
+    std::vector<float> outputpdgammaHostData = {0, 1, 2, 3};
 
     // 创建self aclTensor
     ret = CreateAclTensor(dyHostData, dyShape, &dyDeviceAddr, aclDataType::ACL_FLOAT, &dy);
